Add parameter for whisper API version
Refactor to allow different versions
