+++ test/test_metrics_analyzer.py
@@ -11,7 +11,7 @@ from whisper_automatic_test.metrics_analyzer import MetricsAnalyzer
 from whisper_automatic_test.scenario import Scenario
 from whisper_automatic_test.scenario_reader import get_scenarios_from_csv_file
 from whisper_automatic_test.suggestions_response import SuggestionsResponse
-from whisper_automatic_test.whisper_api_adapter import whisper_response_to_suggestions
+from whisper_automatic_test.whisper_api_adapter.whisper_api_adapter import whisper_response_to_suggestions
 
 SCENARIO_FILE_PATH = 'test/resources/test_scenarios_for_metrics.csv'
 EXAMPLE_WHISPER_RESPONSE_FILE_PATH = 'test/resources/example_whisper_response_3_links.json'
