+++ test/test_suggestions_responses_analyzer.py
@@ -11,7 +11,7 @@ from whisper_automatic_test.scenario_reader import get_scenarios_from_csv_file
 from whisper_automatic_test.suggestion import Suggestion
 from whisper_automatic_test.suggestions_response import SuggestionsResponse
 from whisper_automatic_test.suggestions_responses_analyzer import SuggestionsResponsesAnalyzer
-from whisper_automatic_test.whisper_api_adapter import whisper_response_to_suggestions
+from whisper_automatic_test.whisper_api_adapter.whisper_api_adapter import whisper_response_to_suggestions
 
 SCENARIO_FILE_PATH = 'test/resources/test_scenarios_for_suggestions_responses_analyzer.csv'
 SCENARIOS_STARTING_WITH_SAME_FILE_PATH = 'test/resources/test_scenarios_second_scenario_starts_with_same.csv'
