+++ test/whisper_api_adapter/get_json_for_whisper_api_for_each_version/test_get_json_for_whisper_api_version_10.py
@@ -1,0 +1,28 @@
+import unittest
+
+from whisper_automatic_test.request import Request
+from whisper_automatic_test.whisper_api_adapter.whisper_api_adapter import get_json_for_whisper_api
+
+
+class TestGetJsonVersion10(unittest.TestCase):
+    def test_request_to_data_type_asker(self):
+        request = Request('asker', 'Hello world', 'To ignore', 'To ignore')
+        actual_json = get_json_for_whisper_api('10', request, 'myChatKey')
+        expected_json = {
+            'chatkey': 'myChatKey',
+            'querry': 'Hello world',
+            'query': 'Hello world',
+            'type': 0
+        }
+        self.assertEquals(expected_json, actual_json)
+
+    def test_request_to_data_type_agent(self):
+        request = Request('agent', 'Hello world', 'To ignore', 'To ignore')
+        actual_json = get_json_for_whisper_api('10', request, 'myChatKey')
+        expected_json = {
+            'chatkey': 'myChatKey',
+            'querry': 'Hello world',
+            'query': 'Hello world',
+            'type': 1
+        }
+        self.assertEquals(expected_json, actual_json)
